Verstecken und Anzeigen von Knoten jetzt auch in der Categoryview

+++ org.eclipselabs.recommenders.bookmark/src/org/eclipselabs/recommenders/bookmark/copyCutPaste/CutHandler.java
@@ -48,8 +48,10 @@ public class CutHandler
 
 			manager.getExpandedStorage().removeNode(refNode);
 
+			if (refNode != node) {
+				TreeUtil.unlink(refNode);
+			}
 			TreeUtil.unlink(node);
-			TreeUtil.unlink(refNode);
 		}
 	}
